pyproject.toml
[project]
name = "people-search-bench"
version = "1.0.0"
description = "The first open benchmark for evaluating AI-powered people search agents"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.9",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.8",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.21",
]

[project.urls]
Homepage = "https://github.com/LessieAI/people-search-bench"
